requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
gunicorn>=21.2
psycopg2-binary>=2.9
dj-database-url>=2.1
whitenoise>=6.6
